+++ docs/tasks/2021y-01m-31d-0027-mygtest_2_0_0.md
@@ -1,0 +1,23 @@
+
+| дата начала         |   дата конца        | длительность  | исполнитель  |
+|:-------------------:|:-------------------:|:-------------:|:------------:|
+| 2020y-01m-31d 16:35 | 2020y-01m-31d 16:50 | 15 min        | Kartonagnick |
+
+Проект mygtest  
+--------------
+
+Переименовываем файлы:  
+  - [ ] версия `2.0.0`  
+  - [ ] было:  
+    - [ ] test_list_dev.hpp  
+    - [ ] test_list_stable.hpp  
+  - [ ] стало:  
+    - [ ] test-develop.hpp  
+    - [ ] test-stable.hpp  
+
+Меняем слеши:  
+  - [ ] было:  
+    - [ ] example\additional.method_000(tdd)  
+  - [ ] стало:  
+    - [ ] example/additional.method_000(tdd)  
+
