+++ projects/mygtest/include/mygtest/confbuild.hpp
@@ -39,9 +39,9 @@
     #define dTXT_RELEASE_STABLE "unstable"
 #endif
 
-#define dVERSION_MAJOR      1
+#define dVERSION_MAJOR      2
 #define dVERSION_MINOR      0
-#define dVERSION_PATCH      1
+#define dVERSION_PATCH      0
 
 #define dSTRINGIZE(...) #__VA_ARGS__
 #define dSSTRINGIZE(x) dSTRINGIZE(x)
