+++ projects/mygtest/include/mygtest/macro-overload-7.hpp
