+++ projects/mygtest/include/mygtest/modern.hpp
@@ -1,5 +1,5 @@
 
 #pragma once
-#include <mygtest/test_list.hpp>
+#include <mygtest/test-list.hpp>
 #include <mygtest/extension.hpp>
