+++ projects/mygtest/test/src/test-classic.cpp
@@ -1,5 +1,5 @@
 
-#include <mygtest/test_list.hpp>
+#include <mygtest/test-list.hpp>
 //==============================================================================
 //==============================================================================
